requirements.txt
pandas>=2.1.0
numpy>=1.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
jupyter>=1.0.0
plotly>=5.15.0
scipy>=1.11.0
python-dateutil>=2.8.2
xgboost>=2.0.0
lightgbm>=4.1.0
joblib>=1.3.0
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.15.0